pyproject.toml
[tool.poetry]
name = "FireBase Extract"
version = "0.1.0"
description = "firebase_storage_extracting"
authors = ["Elijah Arbee <ec.arbee1@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
flask = "^3.0.0"
google-auth = "^2.28.1"
xds-protos = "^1.62.0"
firebase-admin = "^6.4.0"
alt-bucket = "^0.0.2"
firestore = "^0.0.8"
agent-admin = "^0.1.1"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"